+++ lib/cached/md5_cached.cmake
@@ -1,4 +1,4 @@
 set(lexer_lxx_md5_cached "29c69cfc368018c0bf8fe3d63f00c677")
-set(parser_yxx_md5_cached "918420c1d377bdd7b1a5b87ede928f59")
+set(parser_yxx_md5_cached "dea770003b555904ec408e964ec9497a")
 set(regex_lexer_lxx_md5_cached "d4bdc40f824511fb088b32f57405136c")
 set(regex_parser_yxx_md5_cached "e44695248dcfad64a56982f126f8e4f5")